// src/features/press-release/types/pressRelease.type.ts
import { CategoryType } from "@/features/article/types/category.type";
import { Meta } from "@/types/meta.type";

export type PressReleaseType = {
  id: number;
  title: string;
  description: string;
  slug: string;
  content: string;
  category_id: string;
  category: CategoryType,
  status: string;
  published_at: string;
  user: {
    name: string
  }
  thumbnail: string;
  meta: Meta[] | [];
  attachments: ({ url: string, original_name: string })[];
  village_id: number,
};

export type ListPressRelease = {
  data: PressReleaseType[];
  meta: Meta
}